2023_11_06_PbodyProject init branch

+++ quantification/measures.py
@@ -229,7 +229,7 @@ def count_rna_close_pbody_global(pbody_label: np.ndarray, spots_coords: 'list[tu
 
     if len(spots_coords) == 0 :
         res = {
-            '{0} {1} nm'.format(spot_type, distance): pd.Series() for distance in distance_nm
+            '{0} {1} nm'.format(spot_type, distance): pd.Series(dtype= float) for distance in distance_nm
         }
         return res
     if len(spots_coords[0]) == 2 :
